refactor(frontend): コンポーネント分割
