+++ frontend/src/routes/admin/submissions/$submissionId/-components/submission-info/index.ts
@@ -1,0 +1,1 @@
+export { SubmissionInfo } from './submission-info';
