+++ frontend/src/routes/admin/submissions/$submissionId/-components/submissioned-code/index.ts
@@ -1,0 +1,1 @@
+export { SubmissionedCode } from './submissioned-code';
