+++ frontend/src/routes/admin/submissions/$submissionId/-components/test-results/index.tsx
@@ -1,0 +1,1 @@
+export { TestResults } from './test-results';
